requirements.txt
passporteye>=2.1.0
easyocr>=1.6.0
matplotlib>=3.5.0
opencv-python>=4.5.0
python-dateutil>=2.8.0
pdf2image>=1.16.0
pandas>=1.3.0
openpyxl>=3.0.0
tqdm>=4.60.0
Pillow>=9.0.0
numpy>=1.21.0
streamlit>=1.25.0
watchdog>=3.0.0
pyngrok
